# server/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: LeadFlow CRM API
  description: API REST para CRM de Leads B2B y Pipeline Comercial con Express 5, TypeScript y MySQL 8.4 LTS.
  version: 1.0.0
  contact:
    name: LeadFlow Engineering
    email: dev@leadflow.io
servers:
  - url: http://localhost:5000/api/v1
    description: Servidor de Desarrollo Local
paths:
...
  /stages:
    get:
      tags:
        - Etapas de Pipeline
      summary: Listar todas las etapas del pipeline comercial
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de etapas ordenadas con conteos de prospectos
    post:
      tags:
        - Etapas de Pipeline
      summary: Crear nueva etapa de pipeline
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStageInput'
      responses:
        '201':
          description: Etapa creada exitosamente
...
  /leads:
    get:
      tags:
        - Prospectos Comerciales (Leads)
      summary: Listar leads con filtros, búsqueda y paginación
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: stage_id
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        '200':
          description: Listado paginado de prospectos
    post:
      tags:
        - Prospectos Comerciales (Leads)
      summary: Crear nuevo prospecto comercial
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeadInput'
      responses:
        '201':
          description: Prospecto creado exitosamente
...
  /activities:
    get:
      tags:
        - Actividades Comerciales
      summary: Listar actividades con filtros
      security:
        - BearerAuth: []
      parameters:
        - name: lead_id
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [call, meeting, email, note]
      responses:
        '200':
          description: Listado de actividades
    post:
      tags:
        - Actividades Comerciales
      summary: Registrar nueva actividad
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityInput'
      responses:
        '201':
          description: Actividad registrada
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateStageInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Negociación Final
        color:
          type: string
          example: '#6366f1'
        order_index:
          type: integer
          example: 5
        is_won:
          type: boolean
          default: false
        is_lost:
          type: boolean
          default: false
...
    CreateLeadInput:
      type: object
      required:
        - company_name
        - contact_name
        - email
        - stage_id
      properties:
        company_name:
          type: string
          example: InnovaTech Solutions S.L.
        contact_name:
          type: string
          example: Sofía Balaguer
        email:
          type: string
          example: sofia.b@innovatech.es
        phone:
          type: string
          example: +34 932 112 344
        stage_id:
          type: integer
          example: 1
        source_id:
          type: integer
          example: 1
        value_amount:
          type: number
          example: 18500
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        notes:
          type: string
          example: Interesados en migración de CRM.
...
    CreateActivityInput:
      type: object
      required:
        - lead_id
        - type
        - summary
      properties:
        lead_id:
          type: integer
          example: 1
        type:
          type: string
          enum: [call, meeting, email, note]
          example: call
        summary:
          type: string
          example: Llamada de presentación de producto
        scheduled_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...